+++ app/Iconsets/Iconsets/Fontawesome.php
@@ -97,11 +97,11 @@ class Fontawesome extends Iconset_Base implements Iconset {
 		foreach ( $this->get_icon_codes() as $local_filetype => $icon ) {
 			list($type, $subtype) = Helper::get_type_and_subtype_from_mimetype( $local_filetype );
 			if ( empty( $types[ $type ] ) ) {
-				$style         .= '.wp-block-downloadlist-list.iconset-' . $term_slug . ' .file_' . $type . ':before { content: "' . $icon . '";font-family: "Font Awesome 6 Free", sans-serif;font-size: ' . $width . 'px;font-weight: 900 }';
+				$style         .= '.wp-block-downloadlist-list.iconset-' . $term_slug . ' .file_' . $type . ':before { content: "' . $icon . '";font-family: "Font Awesome 7 Free", sans-serif;font-size: ' . $width . 'px;font-weight: 900 }';
 				$types[ $type ] = 1;
 			}
 			if ( ! empty( $subtype ) && empty( $types[ $subtype ] ) ) {
-				$style            .= '.wp-block-downloadlist-list.iconset-' . $term_slug . ' .file_' . $subtype . ':before { content: "' . $icon . '";font-family: "Font Awesome 6 Free", sans-serif;font-size: ' . $width . 'px;font-weight: 900 }';
+				$style            .= '.wp-block-downloadlist-list.iconset-' . $term_slug . ' .file_' . $subtype . ':before { content: "' . $icon . '";font-family: "Font Awesome 7 Free", sans-serif;font-size: ' . $width . 'px;font-weight: 900 }';
 				$types[ $subtype ] = 1;
 			}
 		}